The program now accepts an argument in the cmd-line: the account to which to register its performance.
Fixing README to compile with pthread.

The program now puts its working in a worker-function, so its
working can be started in a separate new pthread (although it
uses no "worker" pthread yet.)

It is the embedded mode of the NewRelic SDK the one which starts
the SDK engine in a separate pthread (at least in the NewRelic
Agent SDK version 0.16.1), and tracing reveals that this
NewRelic thread is the one which connects to the NewRelic
collector site ("collector.newrelic.com") and sends the stats
to it.