adding a script to download the NewRelic SDK, no matter if it changes URLs
Eg., three days ago NewRelic released a new version of its Agent SDK (before
was version 0.16.1, now it is 0.16.2). When NewRelic releases a new version
(TAR-ball) it deletes the previous one, so the old download link in this
Makefile would fail.

The old download link was:

http://download.newrelic.com/agent_sdk/nr_agent_sdk-v0.16.1.0-beta.x86_64.tar.gz

that is now un-existant, and which exists is this:

http://download.newrelic.com/agent_sdk/nr_agent_sdk-v0.16.2.0-beta.x86_64.tar.gz

This script

   download_NewRelic_Agent_SDK.sh

avoids that, it adjusts itself to the current URL.

Adjusting the Makefile accordingly.