No need for the wrapper function to our activity to have that signature
The function-signature that it had before was for the very original version
of the program using the New Relic Agent SDK embedded mode in C/C++, where
its documentation indicated that it required pthreads. The C/C++ application
does requires to be linked with the pthreads library because the embedded mode
runs in its own pthread (sending data to https://collector.newrelic.com), so
the main pthread may run directly our task.